requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
opencv-python
Pillow
tqdm
streamlit
joblib
